Added scripts
Added a bunch of scripts:
Tag image is supposed to take a photo, and give it tags and rename the photo based on what is seen.
rename image will take a single photo and rename it based on what is in the photo.
batchrenamephotos will do what rename image does, but instead you pass in a folder of photos for them all to be automatically renamed.
imagenet_labels.txt is a database of labels to give the photos. This is 1000 imagenet labels to train the mobilenet.